产品订单bug 修复

+++ src/main/java/com/std/forum/ao/IProductAO.java
@@ -14,7 +14,7 @@ public interface IProductAO {
 
     public int editProduct(Product data);
 
-    public int editProductStatus(String code, Integer price);
+    public int editProductStatus(String code, Long price);
 
     public Paginable<Product> queryProductPage(int start, int limit,
             Product condition);
